organize phase4 modules
