+++ compare_pca_umap.py
@@ -13,7 +13,7 @@ import seaborn as sns
 import pandas as pd
 import yaml
 
-from phase4_functions import (
+from phase4.functions import (
     load_datasets,
     prepare_data,
     select_variables,
