+++ compare_umap_pacmap.py
@@ -17,7 +17,7 @@ import pandas as pd
 import seaborn as sns
 import yaml
 
-from phase4_functions import (
+from phase4.functions import (
     load_datasets,
     prepare_data,
     select_variables,
