+++ export_mfa_group_contributions.py
@@ -10,7 +10,7 @@ from typing import Any, Mapping
 import pandas as pd
 import yaml
 
-import phase4_functions as pf
+import phase4.functions as pf
 
 
 def load_config(path: Path) -> Mapping[str, Any]:
