+++ phase4/create_cover.py
