+++ tests/test_evaluate_methods.py
@@ -2,7 +2,7 @@ import pandas as pd
 import numpy as np
 import pytest
 
-import phase4_functions as pf
+import phase4.functions as pf
 
 
 def sample_df() -> pd.DataFrame:
