+++ tests/test_load_datasets.py
@@ -1,7 +1,7 @@
 import pandas as pd
 import pytest
 from pathlib import Path
-import phase4_functions as pf
+import phase4.functions as pf
 
 
 def _make_sample_config(tmp_path: Path) -> dict:
