+++ tests/test_nonlin_methods.py
@@ -1,7 +1,7 @@
 import builtins
 import numpy as np
 import pandas as pd
-import phase4_functions as pf
+import phase4.functions as pf
 
 
 def sample_df():
